+++ src/App.tsx
@@ -8,8 +8,8 @@ import AnimalsListPage from "./pages/animalsListPage/animalsListPage"; // page d
 import Signup_assoPage from "./pages/signupPage/signup_assoPage"; // page de l'inscription d'une association
 import Signup_faPage from "./pages/signupPage/signup_faPage"; // page de l'inscription d'une famille
 import HomePage from "./pages/homePage/homePage"; // page d'accueil
-import ProfilePageAsso from "./pages/spaceAsso/AssociationProfile.tsx";
-import ProfilePageFamily from "./pages/spaceFamily/FamilyProfile.tsx";
+import ProfilePageAsso from "./pages/spaceAsso/associationProfile/associationProfile.tsx";
+import ProfilePageFamily from "./pages/spaceFamily/familyProfile/familyProfile.tsx";
 import Header from "./components/header/header";
 import Footer from "./components/footer/footer";
